src/main/java/com/fiap/postech/techchallenge/fastfoodpayment/application/api/config/ControllerExceptionHandler.java
package com.fiap.postech.techchallenge.fastfoodpayment.application.api.config;

import com.fiap.postech.techchallenge.fastfoodpayment.infra.gateway.feign.mercadopago.exception.MercadoPagoQRCodeException;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class ControllerExceptionHandler {


    @ExceptionHandler(value = {MercadoPagoQRCodeException.class})
    protected ResponseEntity<ApiError> handlerMercadoPagoQRCodeException(MercadoPagoQRCodeException ex) {
        ApiError error = new ApiError(ex.getMessage(),HttpStatus.BAD_REQUEST.value());
        return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
    }


}
